Add the database template for MongoDB compass
This is the template for our database using the current schema from the wiki. I made them json files and explained in the readme how to setup the initial database. There was a way to do it with command line but I didn't know if everyone had the path for the commands already in their terminal. But after import and following instructions we should all have the same database structure with example documents.