pom.xml creation
This is the updated pom.xml for running the code with maven. It includes all the required dependencies and versions for Java it needs a release of 17 or higher for the mongo version 5.* if using a later version of java change the poml under properties at the top the compiler release that says 17 so if using 21 change pom to show that.